requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
pinecone-client
pypdf
beautifulsoup4
tiktoken
python-dotenv
langchain_pinecone
langchain-huggingface
sentence-transformers
requests